+++ src/SqlBuildManager.Console.ExternalTest/BatchTests.cs
@@ -1494,7 +1494,7 @@ namespace SqlBuildManager.Console.ExternalTest
         [DataRow("runthreaded", "TestConfig/settingsfile-batch-windows.json")]
         [DataRow("run", "TestConfig/settingsfile-batch-linux.json")]
         [DataTestMethod]
-        public void Batch_Override_DacpacSource_FocceApplyCustom_Success(string batchMethod, string settingsFile)
+        public void Batch_Override_DacpacSource_ForceApplyCustom_Success(string batchMethod, string settingsFile)
         {
             int removeCount = 1;
             string server, database;