+++ src/SqlBuildManager.Console/CommandLine/CommandLineBuilder.ContainerApp.cs
@@ -30,6 +30,7 @@ namespace SqlBuildManager.Console.CommandLine
                 cmd.AddRange(ContainerAppOptions);
                 IdentityArgumentsForContainerApp.ForEach(o => { if (o.Name == "identityname") { o.IsRequired = false; } cmd.Add(o); });
                 cmd.AddRange(ConnectionAndSecretsOptions);
+                cmd.AddRange(EventHubResourceOptions);
                 ////TODO: Enable Managed Identity. For now, ManagedIdentity for SQL Auth is not available on Container Apps...
                 //cmd.AddRange(DatabaseAuthArgs);
                 cmd.Add(usernameOption);
@@ -74,7 +75,7 @@ namespace SqlBuildManager.Console.CommandLine
                 cmd.AddRange(ContainerAppOptions);
                 IdentityArgumentsForContainerApp.ForEach(o => { if (o.Name == "identityname") { o.IsRequired = false; } cmd.Add(o); });
                 cmd.AddRange(ConnectionAndSecretsOptions);
-
+                cmd.AddRange(EventHubResourceOptions);
                 //TODO: Enable Managed Identity. For now, ManagedIdentity for SQL Auth is not available on Container Apps...
                 //cmd.AddRange(DatabaseAuthArgs);
                 cmd.Add(usernameOption);
@@ -164,7 +165,7 @@ namespace SqlBuildManager.Console.CommandLine
                 cmd.AddRange(ContainerAppOptions);
                 IdentityArgumentsForContainerApp.ForEach(o => { if (o.Name == "identityname") { o.IsRequired = false; } cmd.Add(o); });
                 cmd.AddRange(ConnectionAndSecretsOptions);
-
+                cmd.AddRange(EventHubResourceOptions);
                 //TODO: Enable Managed Identity. For now, ManagedIdentity for SQL Auth is not available on Container Apps...
                 //DatabaseAuthArgs.ForEach(o => cmd.Add(o));
                 cmd.Add(usernameOption);
@@ -198,6 +199,7 @@ namespace SqlBuildManager.Console.CommandLine
                 cmd.Add(keyVaultNameOption);
                 cmd.Add(serviceBusconnectionOption);
                 cmd.Add(eventhubconnectionOption);
+                cmd.AddRange(EventHubResourceOptions);
                 cmd.Handler = CommandHandler.Create<CommandLineArgs, bool, DateTime?, bool>(Worker.MonitorContainerAppRuntimeProgress);
 
                 return cmd;